[refactor]componentsフォルダの位置を適切な場所に
