+++ server/front/src/components/Post.vue
